feat: lab2
put, new exception

+++ 1_Spring_Framework_Core_and_DI/src/main/java/com/webdev/sdc/service/CurrencyService.java
@@ -3,11 +3,13 @@ package com.webdev.sdc.service;
 import com.webdev.sdc.dto.CurrencyDto;
 import com.webdev.sdc.exception.ConflictException;
 import com.webdev.sdc.exception.GlobalException;
+import com.webdev.sdc.exception.RandomException;
 import com.webdev.sdc.model.Currency;
 import com.webdev.sdc.model.CurrencyEntity;
 import com.webdev.sdc.repository.CurrencyRepository;
 import org.springframework.stereotype.Service;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @Service
@@ -35,19 +37,40 @@ public class CurrencyService {
             throw new ConflictException("Currency exists");
         }
 
-        CurrencyEntity currecnyEntity = new CurrencyEntity(
+        CurrencyEntity currencyEntity = new CurrencyEntity(
                 null,
                 currencyDto.getType(),
                 currencyDto.getRate()
         );
 
-        return this.repository.save(currecnyEntity);
+        return this.repository.save(currencyEntity);
     }
 
     public void delete(Long id) {
         if (id == null) {
             throw new GlobalException("id is null");
         }
+        if (Math.random() < 0.5) {
+            throw new RandomException();
+        }
         repository.deleteById(id);
+    }
+
+    public CurrencyEntity update(Long id, @Valid CurrencyDto currencyDto) {
+        if (id == null) {
+            throw new GlobalException("id is null");
+        }
+
+        if (currencyDto == null) {
+            throw new GlobalException("currencyDto is null");
+        }
+
+        CurrencyEntity currencyEntity = new CurrencyEntity(
+                id,
+                currencyDto.getType(),
+                currencyDto.getRate()
+        );
+
+        return repository.save(currencyEntity);
     }
 }
